refactor: http-exception filter move to filters folder
